_bib/requirements.txt
pytest
delegator.py
pybtex
